requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
openai==1.6.1
python-dotenv==1.0.0
redis==5.0.1
requests==2.31.0
gunicorn==21.2.0
python-json-logger==2.0.7
flask-cors==4.0.0
flask-limiter==3.5.0
cachetools==5.3.2
pydantic==2.5.2
jinja2==3.1.2
click==8.1.7
itsdangerous==2.1.2
MarkupSafe==2.1.3